# src/main/resources/static/swagger-ui/openapi.yaml
openapi: 3.0.0
info:
  title: API de Residencias
  version: 1.0.0
  description: API para gestionar residencias, con operaciones de registro y consulta.
paths:
...
  /resi/resident/getAll:
    get:
      tags:
        - Residente
      summary: Obtener residentes
      description: Devuelve una lista de residentes.  Y filtrado por los sigueintes parametros. Hay que tener en cuenta sigue una proridad, documento de identidad, fecha de nacimiento y mes/año
      security:
        - bearerAuth: []
      parameters:
        - name: documentoIdentidad
          in: query
          required: false
          schema:
            type: string
        - name: fechaNacimiento
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: year
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: month
          in: query
          required: false
          schema:
              type: integer
              format: int64
        - name: maxAge
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: minAge
          in: query
          required: false
          schema:
              type: integer
              format: int64
        - name: idJuego
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: idEventoSalida
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Residente(s) econtrado(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Residente.ResponseDto'
        '404':
          description: Residencia invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                value:
                  mensaje: Residencia invalida
                  codigo: 1006
                  status: 404
                  timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
        '406':
          description: Rango de edad invalido, el minimo no puede ser mayor que el maximo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RANGO_EDAD_INVALIDO:
                  value:
                    mensaje: Rango de edad invalido
                    codigo: 1029
                    status: 406
                    timestamp: 2025-04-13T21:00:00
  /resi/resident/getAll/bajas:
    get:
      tags:
        - Residente
      summary: Obtener residentes en baja
      description: Devuelve una lista de residentes en baja.
      security:
        - bearerAuth: []
      parameters:
        - name: documentoIdentidad
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Residente(s) econtrado(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Residente.ResponseDto'
        '404':
          description: Residencia invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                value:
                  mensaje: Residencia invalida
                  codigo: 1006
                  status: 404
                  timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
...
components:
  schemas:
...
    Residente.ResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
        nombre:
          type: string
          example: Carlos
        apellido:
          type: string
          example: Pérez
        fechaNacimiento:
          type: string
          format: date
          example: 1945-06-15
        idResidencia:
          type: integer
          format: int64
          example: 1
        documentoIdentidad:
          type: string
          example: Y8440460X
        baja:
          type: boolean
          example: false
...
    Error.ResponseDto:
      type: object
      properties:
        mensaje:
          type: string
        codigo:
          type: integer
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        mensaje: Faltan campos obligatorios
        codigo: 1001
        status: 400
        timestamp: 2025-04-13T21:00:00
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT